pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "fenz050896",email = "michaelfitrahjackson@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "alembic (>=1.15.2,<2.0.0)",
    "blinker (>=1.9.0,<2.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "flask-alembic (>=3.1.1,<4.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "greenlet (>=3.2.1,<4.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "mako (>=1.3.10,<2.0.0)",
    "markupsafe (>=3.0.2,<4.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "tenseal (>=0.3.16,<0.4.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
setuptools = "^80.8.0"
faker = "^37.3.0"
